moved load_config_file to utils
